DF-9778 review fix
Former-commit-id: 49d4cd0d6ffe46a95175ac444385d57c524e5d71

+++ Sources/Internal/Render/Highlevel/Vegetation/VegetationRenderObject.h
@@ -321,7 +321,7 @@ inline uint32 VegetationRenderObject::MapToResolution(float32 squareDistance)
         if (squareDistance > resolutionRanges[i].x &&
             squareDistance <= resolutionRanges[i].y)
         {
-            resolutionId = i;
+            resolutionId = static_cast<uint32>(i);
             break;
         }
     }
